osmBack/pyproject.toml
[project]
name = "osmBack"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Elhussin", email = "hasin3112@gmail.com"},
]
dependencies = [
    "asgiref==3.9.2",
    "gunicorn==21.2.0",
    "attrs==25.3.0",
    "babel==2.17.0",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "cryptography==46.0.1",
    "dj-rest-auth==7.0.1",
    "django-admin-interface==0.30.1",
    "django-allauth==65.11.2",
    "django-colorfield==0.14.0",
    "django-cors-headers==4.9.0",
    "django-extensions==4.1",
    "django-filter==25.1",
    "django-money==3.5.4",
    "django-simple-history==3.10.1",
    "django-tenants==3.9.0",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-spectacular==0.28.0",
    "drf-spectacular-sidecar==2025.9.1",
    "idna==3.10",
    "inflection==0.5.1",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "paypalrestsdk==1.13.3",
    "pillow==11.3.0",
    "psycopg2-binary==2.9.11",
    "py-moneyed==3.0",
    "pycparser==2.23",
    "pyjwt==2.10.1",
    "pyopenssl==25.3.0",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "python-slugify==8.0.4",
    "pyyaml==6.0.3",
    "referencing==0.36.2",
    "requests==2.32.5",
    "rpds-py==0.27.1",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "text-unidecode==1.3",
    "typing-extensions==4.15.0",
    "tzdata==2025.2; sys_platform == \"win32\"",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "whitenoise==6.11.0",
    "python",
    "django==5.2.9",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
